Structure custom exceptions | Alter get_sentence method
